Started moving delegations-related helpers into their own module
