+++ contracts/mixnet/src/mixnodes/mod.rs
@@ -1,5 +1,6 @@
 pub mod bonding_queries;
 pub mod bonding_transactions;
+pub mod delegation_helpers;
 pub mod delegation_queries;
 pub mod delegation_transactions;
 pub mod layer_queries;
